requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anthropic==0.38.0
anyio==4.6.2.post1
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
build==1.2.2.post1
CacheControl==0.14.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
cleo==2.1.0
click==8.1.7
cohere==5.11.3
colorama==0.4.6
commonmark==0.9.1
crashtest==0.4.1
cryptography==44.0.1
dataclasses-json==0.6.7
datasets==2.21.0
denser-retriever @ git+https://github.com/denser-org/denser-retriever.git@74da4a03fbfdd8fa3195e12fb58cae40db5ead31
dill==0.3.8
distlib==0.3.9
distro==1.9.0
dulwich==0.22.7
elastic-transport==8.15.1
elasticsearch==8.15.1
emoji==2.14.1
environs==9.5.0
eval_type_backport==0.2.2
fake-useragent==1.5.1
fastavro==1.9.7
fastjsonschema==2.21.1
filelock==3.16.1
filetype==1.2.0
frozenlist==1.5.0
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.38.0
google-genai==1.2.0
greenlet==3.1.1
grpcio==1.67.1
grpcio-tools==1.67.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
hyperframe==6.0.1
idna==3.10
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.4
jiter==0.7.0
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==25.6.0
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.10
langchain-elasticsearch==0.2.2
langchain-milvus==0.1.1
langchain-qdrant==0.1.4
langchain-text-splitters==0.2.2
langdetect==1.0.9
langsmith==0.1.139
lxml==5.3.1
MarkupSafe==3.0.2
marshmallow==3.23.1
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.13.2
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
olefile==0.47
openai==1.62.0
orjson==3.10.11
packaging==24.1
pandas==2.2.3
parameterized==0.9.0
pillow==10.4.0
pkginfo==1.12.0
platformdirs==4.3.6
poetry==2.0.1
poetry-core==2.0.1
portalocker==2.10.1
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.1
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.6.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.18.0
pymilvus==2.4.9
PyMuPDF==1.24.13
pypdf==4.3.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2025.2.8
python-magic==0.4.27
python-oxmsg==0.0.2
pytrec_eval==0.5
pytz==2024.2
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
qdrant-client==1.12.1
RapidFuzz==3.12.1
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==10.16.2
rpds-py==0.20.1
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.2.1
setuptools==75.3.0
shellingham==1.5.4
simsimd==6.0.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
streamlit==1.39.0
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.20.3
toml==0.10.2
tomlkit==0.13.2
torch==2.5.1
tornado==6.4.1
tqdm==4.66.6
transformers==4.46.2
trove-classifiers==2025.1.15.22
typer==0.12.5
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
unstructured==0.16.20
unstructured-client==0.29.0
urllib3==2.2.3
virtualenv==20.29.2
watchdog==5.0.3
webencodings==0.5.1
websockets==14.2
wrapt==1.17.2
xgboost==2.1.2
xxhash==3.5.0
yarl==1.17.1
